+++ dev/UltimaXNA.csproj
@@ -243,6 +243,7 @@
     <Compile Include="Core\Configuration\SettingsFile.cs" />
     <Compile Include="Ultima\Data\Body.cs" />
     <Compile Include="Ultima\Data\ContentItem.cs" />
+    <Compile Include="Ultima\Data\ContextMenuItem.cs" />
     <Compile Include="Ultima\Data\HouseRevisionState.cs" />
     <Compile Include="Ultima\Data\Hues.cs" />
     <Compile Include="Ultima\Data\Magery.cs" />
@@ -395,7 +396,7 @@
     <Compile Include="Ultima\Network\Client\UnicodeSpeechPacket.cs" />
     <Compile Include="Ultima\Network\Client\UOGInfoRequestPacket.cs" />
     <Compile Include="Ultima\World\WorldClient.cs" />
-    <Compile Include="Ultima\Data\ContextMenu.cs" />
+    <Compile Include="Ultima\Data\ContextMenuData.cs" />
     <Compile Include="Ultima\Network\Server\CharacterListUpdatePacket.cs" />
     <Compile Include="Ultima\Network\Incomplete\DisplayGumpFastPacket.cs" />
     <Compile Include="Ultima\Network\Incomplete\DisplayMenuPacket.cs" />
